model update

+++ src/main/java/org/edgegallery/mecm/inventory/service/TenantServiceImpl.java
@@ -65,7 +65,8 @@ class TenantServiceImpl implements TenantService {
         }
         Tenant ot = record.get();
         subtractCount(ot, type);
-        if (ot.getAppLcms() == 0 && ot.getAppStores() == 0 && ot.getMecHosts() == 0) {
+        if (ot.getAppLcms() == 0 && ot.getAppStores() == 0 && ot.getMecHosts() == 0
+                && ot.getMecHwCapabilities() == 0) {
             repository.deleteById(id);
             LOGGER.info("Tenant record with tenant identifier {} deleted", id);
         } else {
@@ -88,7 +89,8 @@ class TenantServiceImpl implements TenantService {
         }
         Tenant ot = record.get();
         clearCountNumber(ot, type);
-        if (ot.getAppLcms() == 0 && ot.getAppStores() == 0 && ot.getMecHosts() == 0) {
+        if (ot.getAppLcms() == 0 && ot.getAppStores() == 0 && ot.getMecHosts() == 0
+                && ot.getMecHwCapabilities() == 0) {
             repository.deleteById(id);
             LOGGER.info("Tenant record with tenant identifier {} deleted", id);
         } else {
@@ -107,6 +109,9 @@ class TenantServiceImpl implements TenantService {
                 break;
             case MEC_HOST:
                 ot.setMecHosts(0);
+                break;
+            case MEC_HW_CAPABILITY:
+                ot.setMecHwCapabilities(0);
                 break;
             default:
                 LOGGER.error(Constants.INVALID_MODEL_TYPE);
@@ -133,6 +138,13 @@ class TenantServiceImpl implements TenantService {
             case MEC_HOST:
                 if (isNotOverflow(t.getMecHosts(), 1)) {
                     t.setMecHosts(t.getMecHosts() + 1);
+                } else {
+                    LOGGER.error(Constants.VAR_OVERFLOW_ERROR);
+                }
+                break;
+            case MEC_HW_CAPABILITY:
+                if (isNotOverflow(t.getMecHwCapabilities(), 1)) {
+                    t.setMecHwCapabilities(t.getMecHwCapabilities() + 1);
                 } else {
                     LOGGER.error(Constants.VAR_OVERFLOW_ERROR);
                 }
@@ -168,6 +180,13 @@ class TenantServiceImpl implements TenantService {
                     t.setMecHosts(t.getMecHosts() - 1);
                 } else {
                     LOGGER.error("{} for mechost count", Constants.VAR_UNDERFLOW_ERROR);
+                }
+                break;
+            case MEC_HW_CAPABILITY:
+                if (isNotUnderflow(t.getMecHwCapabilities(), 1)) {
+                    t.setMecHwCapabilities(t.getMecHwCapabilities() - 1);
+                } else {
+                    LOGGER.error("{} for mec hardware capability count", Constants.VAR_UNDERFLOW_ERROR);
                 }
                 break;
             default:
